Committed by Vinod Kumar Siripuram(19March2018)
UOB-282
UAT_UOB-Advisor Dashboard-AUM value:

Also added S symbol in front of $ value. But issue is as we are using same dashboard of Advisor/Admin for BB and TCM also it is showing S there also but they are US products so we S would not be required there.----=Need to discuss this point with Prashant.


As Prathemesh disscussed with Prashant, S symbol should be removed from the AUM amount of Advisor Dashboard screen.